wait for expected text
