--- exo-create/test_helpers/empty_directory.go
@@ -1,12 +1,0 @@
-package testHelpers
-
-import (
-	"os"
-)
-
-func EmptyDir(dir string) error {
-	if err := os.RemoveAll(dir); err != nil {
-		return err
-	}
-	return os.Mkdir(dir, os.FileMode(0777))
-}
